+++ packages/wallet/src/Wallet.test.ts
@@ -33,6 +33,10 @@ async function setupWallet(): Promise<Wallet> {
         clientId: 'test',
       },
       networkController: {
+        analyticsOptions: {
+          isRpcEndpointUrlPublic: (): boolean => false,
+          rpcServiceEventsSampleRate: 0,
+        },
         infuraProjectId: 'fake-infura-project-id',
       },
       storageService: {
@@ -99,6 +103,10 @@ describe('Wallet', () => {
           encryptor: new MockEncryptor(),
         },
         networkController: {
+          analyticsOptions: {
+            isRpcEndpointUrlPublic: (): boolean => false,
+            rpcServiceEventsSampleRate: 0,
+          },
           infuraProjectId: 'fake-infura-project-id',
         },
         storageService: {
@@ -151,6 +159,10 @@ describe('Wallet', () => {
           clientId: 'test',
         },
         networkController: {
+          analyticsOptions: {
+            isRpcEndpointUrlPublic: (): boolean => false,
+            rpcServiceEventsSampleRate: 0,
+          },
           infuraProjectId: 'fake-infura-project-id',
         },
         storageService: {
@@ -197,6 +209,10 @@ describe('Wallet', () => {
           clientId: 'test',
         },
         networkController: {
+          analyticsOptions: {
+            isRpcEndpointUrlPublic: (): boolean => false,
+            rpcServiceEventsSampleRate: 0,
+          },
           infuraProjectId: 'fake-infura-project-id',
         },
         storageService: {
@@ -310,6 +326,10 @@ describe('Wallet', () => {
             clientId: 'test',
           },
           networkController: {
+            analyticsOptions: {
+              isRpcEndpointUrlPublic: (): boolean => false,
+              rpcServiceEventsSampleRate: 0,
+            },
             infuraProjectId: 'fake-infura-project-id',
           },
           storageService: {
@@ -345,6 +365,10 @@ describe('Wallet', () => {
             clientId: 'test',
           },
           networkController: {
+            analyticsOptions: {
+              isRpcEndpointUrlPublic: (): boolean => false,
+              rpcServiceEventsSampleRate: 0,
+            },
             infuraProjectId: 'fake-infura-project-id',
           },
           storageService: {
@@ -388,6 +412,10 @@ describe('Wallet', () => {
             clientId: 'test',
           },
           networkController: {
+            analyticsOptions: {
+              isRpcEndpointUrlPublic: (): boolean => false,
+              rpcServiceEventsSampleRate: 0,
+            },
             infuraProjectId: 'fake-infura-project-id',
           },
           storageService: {
@@ -479,6 +507,10 @@ describe('Wallet', () => {
             clientId: 'test',
           },
           networkController: {
+            analyticsOptions: {
+              isRpcEndpointUrlPublic: (): boolean => false,
+              rpcServiceEventsSampleRate: 0,
+            },
             infuraProjectId: 'fake-infura-project-id',
           },
           keyringController: { encryptor: new MockEncryptor() },