+++ packages/wallet/src/initialization/instances/transaction-controller/transaction-controller.test.ts
@@ -119,6 +119,10 @@ function getInstanceOptions(): WalletOptions['instanceOptions'] {
       clientId: 'test',
     },
     networkController: {
+      analyticsOptions: {
+        isRpcEndpointUrlPublic: (): boolean => false,
+        rpcServiceEventsSampleRate: 0,
+      },
       infuraProjectId: 'test-infura-project-id',
     },
     storageService: {